+++ pl_bolts/callbacks/__init__.py
@@ -6,11 +6,13 @@ from pl_bolts.callbacks.data_monitor import ModuleDataMonitor, TrainingDataMonit
 from pl_bolts.callbacks.printing import PrintTableMetricsCallback  # noqa: F401
 from pl_bolts.callbacks.ssl_online import SSLOnlineEvaluator  # noqa: F401
 from pl_bolts.callbacks.variational import LatentDimInterpolator  # noqa: F401
+from pl_bolts.callbacks.verification.batch_gradient import BatchGradientVerificationCallback  # type: ignore
 from pl_bolts.callbacks.vision.confused_logit import ConfusedLogitCallback  # noqa: F401
 from pl_bolts.callbacks.vision.image_generation import TensorboardGenerativeModelImageSampler  # noqa: F401
 from pl_bolts.callbacks.vision.sr_image_logger import SRImageLoggerCallback  # noqa: F401
 
 __all__ = [
+    "BatchGradientVerificationCallback",
     "BYOLMAWeightUpdate",
     "ModuleDataMonitor",
     "TrainingDataMonitor",