+++ pl_bolts/datamodules/vocdetection_datamodule.py
@@ -6,10 +6,10 @@ from pl_bolts.utils import _TORCHVISION_AVAILABLE
 from pl_bolts.utils.warnings import warn_missing_pkg
 
 if _TORCHVISION_AVAILABLE:
-    import torchvision.transforms as T
+    from torchvision import transforms as transform_lib
     from torchvision.datasets import VOCDetection
-else:
-    warn_missing_pkg('torchvision')  # pragma: no-cover
+else:  # pragma: no cover
+    warn_missing_pkg('torchvision')
 
 
 class Compose(object):
@@ -17,12 +17,15 @@ class Compose(object):
     Like `torchvision.transforms.compose` but works for (image, target)
     """
 
-    def __init__(self, transforms):
+    def __init__(self, transforms, image_transforms=None):
         self.transforms = transforms
+        self.image_transforms = image_transforms
 
     def __call__(self, image, target):
         for t in self.transforms:
             image, target = t(image, target)
+        if self.image_transforms:
+            image = self.image_transforms(image)
         return image, target
 
 
@@ -116,8 +119,8 @@ class VOCDetectionDataModule(LightningDataModule):
         *args,
         **kwargs,
     ):
-        if not _TORCHVISION_AVAILABLE:
-            raise ModuleNotFoundError(  # pragma: no-cover
+        if not _TORCHVISION_AVAILABLE:  # pragma: no cover
+            raise ModuleNotFoundError(
                 'You want to use VOC dataset loaded from `torchvision` which is not installed yet.'
             )
 
@@ -146,7 +149,7 @@ class VOCDetectionDataModule(LightningDataModule):
         VOCDetection(self.data_dir, year=self.year, image_set="train", download=True)
         VOCDetection(self.data_dir, year=self.year, image_set="val", download=True)
 
-    def train_dataloader(self, batch_size=1, transforms=None):
+    def train_dataloader(self, batch_size=1, image_transforms=None):
         """
         VOCDetection train set uses the `train` subset
 
@@ -154,12 +157,9 @@ class VOCDetectionDataModule(LightningDataModule):
             batch_size: size of batch
             transforms: custom transforms
         """
-        t = [_prepare_voc_instance]
-        transforms = transforms or self.train_transforms or self._default_transforms()
-        if transforms is not None:
-            t.append(transforms)
-        transforms = Compose(t)
-
+        transforms = [_prepare_voc_instance]
+        image_transforms = image_transforms or self.train_transforms or self._default_transforms()
+        transforms = Compose(transforms, image_transforms)
         dataset = VOCDetection(self.data_dir, year=self.year, image_set="train", transforms=transforms)
         loader = DataLoader(
             dataset,
@@ -172,7 +172,7 @@ class VOCDetectionDataModule(LightningDataModule):
         )
         return loader
 
-    def val_dataloader(self, batch_size=1, transforms=None):
+    def val_dataloader(self, batch_size=1, image_transforms=None):
         """
         VOCDetection val set uses the `val` subset
 
@@ -180,11 +180,9 @@ class VOCDetectionDataModule(LightningDataModule):
             batch_size: size of batch
             transforms: custom transforms
         """
-        t = [_prepare_voc_instance]
-        transforms = transforms or self.val_transforms or self._default_transforms()
-        if transforms is not None:
-            t.append(transforms)
-        transforms = Compose(t)
+        transforms = [_prepare_voc_instance]
+        image_transforms = image_transforms or self.train_transforms or self._default_transforms()
+        transforms = Compose(transforms, image_transforms)
         dataset = VOCDetection(self.data_dir, year=self.year, image_set="val", transforms=transforms)
         loader = DataLoader(
             dataset,
@@ -199,13 +197,10 @@ class VOCDetectionDataModule(LightningDataModule):
 
     def _default_transforms(self):
         if self.normalize:
-            return (
-                lambda image, target: (
-                    T.Compose([
-                        T.ToTensor(),
-                        T.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
-                    ])(image),
-                    target,
-                ),
-            )
-        return lambda image, target: (T.ToTensor()(image), target)
+            voc_transforms = transform_lib.Compose([
+                transform_lib.ToTensor(),
+                transform_lib.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
+            ])
+        else:
+            voc_transforms = transform_lib.Compose([transform_lib.ToTensor()])
+        return voc_transforms