+++ pl_bolts/datasets/mnist_dataset.py
@@ -4,16 +4,16 @@ from pl_bolts.datasets.sr_dataset_mixin import SRDatasetMixin
 from pl_bolts.utils import _PIL_AVAILABLE, _TORCHVISION_AVAILABLE
 from pl_bolts.utils.warnings import warn_missing_pkg
 
-if _PIL_AVAILABLE:
-    import PIL
-else:
-    warn_missing_pkg("PIL", pypi_name="Pillow")  # pragma: no-cover
-
 if _TORCHVISION_AVAILABLE:
     from torchvision.datasets import MNIST
-else:
-    warn_missing_pkg("torchvision")  # pragma: no-cover
+else:  # pragma: no cover
+    warn_missing_pkg('torchvision')
     MNIST = object
+
+if _PIL_AVAILABLE:
+    from PIL import Image
+else:  # pragma: no cover
+    warn_missing_pkg('PIL', pypi_name='Pillow')
 
 
 class BinaryMNIST(MNIST):
@@ -25,6 +25,9 @@ class BinaryMNIST(MNIST):
         Returns:
             tuple: (image, target) where target is index of the target class.
         """
+        if not _TORCHVISION_AVAILABLE:  # pragma: no cover
+            raise ModuleNotFoundError('You want to use `torchvision` which is not installed yet.')
+
         img, target = self.data[idx], int(self.targets[idx])
 
         # doing this so that it is consistent with all other datasets