+++ pl_bolts/models/detection/__init__.py
@@ -1,8 +1,5 @@
-__all__ = []
-
 try:
+    from pl_bolts.models.detection import components  # noqa: F401
     from pl_bolts.models.detection.faster_rcnn import FasterRCNN  # noqa: F401
 except ModuleNotFoundError:  # pragma: no-cover
     pass  # pragma: no-cover
-else:
-    __all__.append('FasterRCNN')