+++ pl_bolts/models/gans/__init__.py
@@ -1,9 +1,11 @@
-from pl_bolts.models.gans.basic.basic_gan_module import GAN
-from pl_bolts.models.gans.srgan.srgan_module import SRGAN
-from pl_bolts.models.gans.srgan.srresnet_module import SRResNet
+from pl_bolts.models.gans.basic.basic_gan_module import GAN  # noqa: F401
+from pl_bolts.models.gans.dcgan.dcgan_module import DCGAN  # noqa: F401
+from pl_bolts.models.gans.srgan.srgan_module import SRGAN  # noqa: F401
+from pl_bolts.models.gans.srgan.srresnet_module import SRResNet  # noqa: F401
 
 __all__ = [
     "GAN",
+    "DCGAN",
     "SRGAN",
     "SRResNet",
 ]