+++ pl_bolts/models/gans/basic/components.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch
-import torch.nn.functional as F
 from torch import nn
+from torch.nn import functional as F
 
 
 class Generator(nn.Module):