+++ pl_bolts/models/self_supervised/amdim/networks.py
@@ -2,8 +2,8 @@ import math
 
 import numpy as np
 import torch
-import torch.nn.functional as F
 from torch import nn
+from torch.nn import functional as F
 
 
 class AMDIMEncoder(nn.Module):