+++ pl_bolts/models/self_supervised/cpc/networks.py
@@ -1,5 +1,5 @@
-import torch.nn.functional as F
 from torch import nn
+from torch.nn import functional as F
 
 
 class CPCResNet(nn.Module):