+++ pl_bolts/models/self_supervised/ssl_finetuner.py
@@ -2,8 +2,8 @@ from typing import List, Optional
 
 import pytorch_lightning as pl
 import torch
-import torch.nn.functional as F
 from pytorch_lightning.metrics import Accuracy
+from torch.nn import functional as F
 
 from pl_bolts.models.self_supervised import SSLEvaluator