+++ tests/models/test_vision.py
@@ -70,7 +70,7 @@ def test_semantic_segmentation():
     model = SemSegment(num_classes=19)
 
     trainer = pl.Trainer(fast_dev_run=True, max_epochs=1)
-    trainer.fit(model, dm)
+    trainer.fit(model, datamodule=dm)
     loss = trainer.progress_bar_dict['loss']
 
     assert float(loss) > 0